refactor: break summary_cache dependency cycle
Move the neutral token-merge helpers to utils and the aggregated-row DTO
to models, and add ExtensionType::scan_rank, so summary_cache no longer
imports from usage or analysis.
