+++ src/utils/mod.rs
@@ -14,6 +14,7 @@ pub mod heap;
 pub mod paths;
 pub mod time;
 pub mod token_extractor;
+pub mod token_merge;
 pub mod usage_processor;
 
 // Public API exports (commonly used across modules)
@@ -42,6 +43,8 @@ pub use paths::{
 };
 pub use time::{now_rfc3339_utc_nanos, parse_iso_timestamp};
 pub use token_extractor::{TokenCounts, extract_token_counts};
+pub(crate) use token_merge::merge_usage_values;
+pub use token_merge::normalize_usage_value;
 pub use usage_processor::{
     CodexTokenTotals, accumulate_i64_fields, accumulate_nested_object, claude_request_context,
     process_claude_usage, process_codex_usage, process_gemini_usage,
